requirements.txt
gspread_dataframe
gspread~=5.4.0
pandas~=1.4.2
requests~=2.28.0
loguru~=0.6.0
oauth2client~=4.1.3
pydantic~=1.9.1
fastapi[all]~=0.78.0
lxml
tld~=0.12.6
SQLAlchemy~=1.4.37
psycopg2-binary
python-dotenv~=0.20.0
beautifulsoup4~=4.11.1
asyncpg